requirement.txt
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==2.2.2
django-allauth==0.39.1
django-appconf==1.0.3
django-autoslug==1.9.6
django-braces==1.13.0
django-cors-headers==3.1.0
django-cors-middleware==1.4.0
django-extensions==2.2.5
django-imagekit==4.0.2
django-model-utils==3.2.0
django-oauth-toolkit==1.2.0
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
drf-yasg==1.17.0
graphviz==0.13
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.17
mysqlclient==1.4.2
oauthlib==3.0.1
packaging==19.2
pilkit==2.0
Pillow==6.1.0
protobuf==3.9.1
pydotplus==2.0.2
PyJWT==1.7.1
pyparsing==2.4.2
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
setuptools==40.8.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3